Fixed bug: wrong game name display
