+++ src/components/PageLayout.jsx
@@ -13,10 +13,8 @@ const PageLayout = ({ children }) => {
         <div className={`font-montserrat relative`}>
             <Header />
             {children}
-            <Footer />
         </div>
     );
 }
 
 export default PageLayout;
-
